Fix calculation of endaddress in cppmod ordt_addr_elem_array
When constructed, the startaddress and endaddress of an ordt_addr_elem_array
instance are the start and end addresses of a single element. In the
initializer of the ordt_addr_elem base class, the endaddress of the whole
array was calculated by adding the stride times the number of reps to the
endaddress passed in. This caused the endaddress of the array to be 1 element
beyond the end of the array, since the first element was already accounted
for. The fix is to add the stride times the reps minus 1.