added use_module_path_defines uvmregs ouput parameter
if false, auto-generated calls to set_rdl_address_map_hdl_path using
path defines will not be included in the model - this is desirable if
multiple instances of a uvm_regs model with different pio rtl are used
in the same bench.  Calls to set_rdl_address_map_hdl_path for each
address map block will need to be performed by user to use hdl paths.